adding utils/startup_utils.py to refine handling of startup in automation
